# Reduce allocation and unnecessary layers