--- a/byteutils/byteutils_test.go
+++ b/byteutils/byteutils_test.go
@@ -32,9 +32,10 @@ func TestReplace(t *testing.T) {
 }
 
 func BenchmarkStringtoSlice(b *testing.B) {
-	b.StopTimer()
-	buf := make([]byte, b.N)
-	b.StartTimer()
-	s := new(string)
-	SliceToString(&buf, s)
+	var s string
+	var buf [1 << 20]byte
+	for i := 0; i < b.N; i++ {
+		s = SliceToString(buf[:])
+	}
+	_ = s // avoid gc to optimize away the loop body
 }