--- a/output_dummy.go
+++ b/output_dummy.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"fmt"
+	"os"
 )
 
 // DummyOutput used for debugging, prints all incoming requests
@@ -16,9 +16,7 @@ func NewDummyOutput() (di *DummyOutput) {
 }
 
 func (i *DummyOutput) Write(data []byte) (int, error) {
-	fmt.Println(string(data))
-
-	return len(data), nil
+	return os.Stdout.Write(data)
 }
 
 func (i *DummyOutput) String() string {