--- a/tcp/tcp_packet.go
+++ b/tcp/tcp_packet.go
@@ -21,8 +21,11 @@ type Packet struct {
 	gopacket.LinkLayer
 
 	// IP Header
-	gopacket.NetworkLayer
 	Version uint8 // Ip version
+	SrcIP   net.IP
+	DstIP   net.IP
+	IHL     uint8
+	Length  uint16
 
 	// TCP Segment Header
 	*layers.TCP
@@ -35,10 +38,12 @@ type Packet struct {
 // ParsePacket parse raw packets
 func ParsePacket(packet gopacket.Packet) (pckt *Packet, err error) {
 	// early check of error
+	if packet == nil {
+		return
+	}
 	defer func() {
 		if packet.ErrorLayer() != nil {
 			err = packet.ErrorLayer().Error()
-			println(err.Error())
 			return
 		}
 	}()
@@ -55,11 +60,17 @@ func ParsePacket(packet gopacket.Packet) (pckt *Packet, err error) {
 
 	// parsing network layer
 	if net4, ok := packet.NetworkLayer().(*layers.IPv4); ok {
-		pckt.NetworkLayer = net4
 		pckt.Version = 4
+		pckt.SrcIP = net4.SrcIP
+		pckt.DstIP = net4.DstIP
+		pckt.IHL = net4.IHL * 4
+		pckt.Length = net4.Length
 	} else if net6, ok := packet.NetworkLayer().(*layers.IPv6); ok {
-		pckt.NetworkLayer = net6
 		pckt.Version = 6
+		pckt.SrcIP = net6.SrcIP
+		pckt.DstIP = net6.DstIP
+		pckt.IHL = 40
+		pckt.Length = net6.Length
 	} else {
 		pckt = nil
 		return
@@ -75,56 +86,23 @@ func ParsePacket(packet gopacket.Packet) (pckt *Packet, err error) {
 	pckt.DataOffset *= 4
 
 	// calculating lost data
-	headerSize := int(uint32(pckt.DataOffset) + uint32(pckt.IHL()))
+	headerSize := int(uint32(pckt.DataOffset) + uint32(pckt.IHL))
 	if pckt.Version == 6 {
 		headerSize -= 40 // in ipv6 the length of payload doesn't include the IPheader size
 	}
-	pckt.Lost = pckt.Length() - uint16(headerSize+len(pckt.Payload))
+	pckt.Lost = pckt.Length - uint16(headerSize+len(pckt.Payload))
 
 	return
 }
 
 // Src returns the source socket of a packet
 func (pckt *Packet) Src() string {
-	return fmt.Sprintf("%s:%d", pckt.SrcIP(), pckt.SrcPort)
+	return fmt.Sprintf("%s:%d", pckt.SrcIP, pckt.SrcPort)
 }
 
 // Dst returns destination socket
 func (pckt *Packet) Dst() string {
-	return fmt.Sprintf("%s:%d", pckt.DstIP(), pckt.DstPort)
-}
-
-// SrcIP returns source IP address
-func (pckt *Packet) SrcIP() net.IP {
-	if pckt.Version == 4 {
-		return pckt.NetworkLayer.(*layers.IPv4).SrcIP
-	}
-	return pckt.NetworkLayer.(*layers.IPv6).SrcIP
-}
-
-// DstIP returns destination IP address
-func (pckt *Packet) DstIP() net.IP {
-	if pckt.Version == 4 {
-		return pckt.NetworkLayer.(*layers.IPv4).DstIP
-	}
-	return pckt.NetworkLayer.(*layers.IPv6).DstIP
-}
-
-// IHL returns IP header length in bytes
-func (pckt *Packet) IHL() uint8 {
-	if l, ok := pckt.NetworkLayer.(*layers.IPv4); ok {
-		return l.IHL * 4
-	}
-	// on IPV6 it's constant, https://en.wikipedia.org/wiki/IPv6_packet#Fixed_header
-	return 40
-}
-
-// Length returns the total length of the packet(IP header, TCP header and the actual data)
-func (pckt *Packet) Length() uint16 {
-	if l, ok := pckt.NetworkLayer.(*layers.IPv4); ok {
-		return l.Length
-	}
-	return pckt.NetworkLayer.(*layers.IPv6).Length
+	return fmt.Sprintf("%s:%d", pckt.DstIP, pckt.DstPort)
 }
 
 // SYNOptions returns MSS and windowscale of syn packets
@@ -205,8 +183,8 @@ Lost Data: %d`,
 		pckt.LinkInfo(),
 		pckt.Src(),
 		pckt.Dst(),
-		pckt.IHL(),
-		pckt.Length(),
+		pckt.IHL,
+		pckt.Length,
 		pckt.Seq,
 		pckt.Ack,
 		pckt.DataOffset,